+++ database/migrations/2021_05_28_091007_pengeluarans.php
@@ -23,9 +23,9 @@ class Pengeluarans extends Migration
             $table->date('invoice');
             $table->date('packing_list');
             $table->string('valuta');
-            $table->integer('kurs');
-            $table->integer('nilai_cif');
-            $table->integer('nilai_barang');
+            $table->float('kurs', 25);
+            $table->float('nilai_cif', 25);
+            $table->float('nilai_barang', 25);
             $table->string('barang');
             $table->date('get_out_start');
             $table->date('get_out_finish');
@@ -33,9 +33,9 @@ class Pengeluarans extends Migration
             $table->integer('jumlah_kemasan');
             $table->string('jenis_kemasan');
             $table->string('merk_kemasan');
-            $table->integer('bruto');
-            $table->integer('netto');
-            $table->integer('volume');
+            $table->float('bruto', 25);
+            $table->float('netto', 25);
+            $table->float('volume', 25);
             $table->softDeletes();
             $table->timestamps();
         });
